refactor `fromCString` to use `size_t` for length

+++ bytecode-vm-compiler/src/compiler.h
@@ -5,4 +5,4 @@
 #include <stdbool.h>
 bool compile(char const* source, Chunk* chunk);
 void setCompilingChunk(Chunk* chunk);
-LoxString* fromCString(char const* cString, int length);
+LoxString* fromCString(char const* cString, size_t length);
